feat(general config): Implementation
- Added org unit selection by group
 - Refactored components to individual files
