+++ src/modules/Config/components/General/components/OrganisationUnitConfig/index.ts
@@ -1,0 +1,1 @@
+export * from "./OrganisationUnitConfig"
